pyproject.toml
[project]
name = "autoco-cos"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "PyYAML>=6.0",
  "PyGithub>=2.3.0",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
